requirements.txt
msal
requests
python-dotenv
python-docx
